+++ src/commands/down.ts
@@ -1,6 +1,7 @@
-import { stopRouterStack } from "../core/router";
+import { clearActiveTcpProtocols, stopRouterStack } from "../core/router";
 
 export async function runDownCommand(): Promise<void> {
   stopRouterStack();
+  clearActiveTcpProtocols();
   process.stdout.write("devrouter is down.\n");
 }